Use tagger instead of morphology analysis
In the old way, we analyzed each word form separately and then tried
to find any analysed lemma. Now we analyze the whole sentence in one
go, which results in better (more precise) analysis.

Also, there was a bug where the technical lemma was used instead of
the raw lemma, preventing us from detecting many boundaries, because
the techlemma was not found in the dictionary. This was fixed as well.