By default, segment text using only our segmenter
The aim of the project has changed. Previously, we were trying
to compare three different segmenters; now, we focus mostly on
developing our own system and keep the other two only for reference
and comparison.

Therefore, restructure the Makefile to reflect this – the main target
`all` now only uses the DeriNet-based segmenter to process text.
The previous target (using several systems) is available as `compare`.