Download the unmodified DeriNet DB from ÚFAL
Previously, we downloaded the DeriNet database from a private server.
Apart from being arguably bad style, it also created a bug when the
content of this server changed. The files there gained an additional
column containing an unexpected floating-point value. This confused
the Segmentace parser and made it crash.

The bug is fixed by downloading the official, unmodified version from
the main repository at ÚFAL.

Fixes #6.