Take the parent segmentation into account
When estimating segmentation quality in EM, the stem mapping algorithm
needs to take the parent word into account. So far, this was done
by finding the best stem – but its affixes were not considered, so
a wrong stem was often found. Fix this by iterating over possible
parent segmentations in the main loop as well, checking the parent
affixes and their compatibility with child affixes in addition to
checking the stem mapping.